scripts/requirements.txt
python-dotenv>=0.19.0
requests>=2.26.0
validators>=0.18.2
psutil>=5.8.0
urllib3>=1.26.7
typing-extensions>=4.0.0
pytest>=6.2.5
pytest-mock>=3.6.1 